Refactored to split queries and models into separate files and folders organized by their respective collection
